Debug

+++ src/utils.h
@@ -5,7 +5,7 @@
  */
 
 // Prototypes
-char *memstr(char *haystack, char *needle, int size);
+char *memstr(char *haystack, char *needle, int64_t size);
 char *from_ngx_str(ngx_pool_t *pool, ngx_str_t ngx_str);
 char *from_ngx_str_malloc(ngx_pool_t *pool, ngx_str_t ngx_str);
 ngx_int_t set_metadata_field (ngx_http_request_t *r, char **field, char *field_name, const char *value);
